fix: pr suggestion about unifying httpx auth tokens
