+++ craft_store/__init__.py
@@ -21,10 +21,9 @@ __version__ = "3.0.0"
 
 
 from . import creds, endpoints, errors, models, publishergateway
+from ._httpx_auth import CandidAuth, DeveloperTokenAuth
 from .auth import Auth
 from .base_client import BaseClient
-from .candidauth import CandidAuth
-from .developer_token_auth import DeveloperTokenAuth
 from .http_client import HTTPClient
 from .store_client import StoreClient
 from .ubuntu_one_store_client import UbuntuOneStoreClient
